+++ Tests/Unit/Gateway/Request/MpDeviceSessionIdTest.php
@@ -5,22 +5,26 @@ namespace MercadoPago\AdbPayment\Tests\Unit\Gateway\Request;
 use PHPUnit\Framework\TestCase;
 use MercadoPago\AdbPayment\Gateway\Request\MpDeviceSessionId;
 use Magento\Payment\Gateway\Data\PaymentDataObject;
-use Magento\Sales\Model\Order\Payment\Interceptor;
+use Magento\Payment\Model\InfoInterface;
 
-class MpDeviceSessionIdTest extends TestCase {
+class MpDeviceSessionIdTest extends TestCase
+{
+    /**
+     * @return void
+     */
     public function testBuildWithSession()
     {
-        $paymentInterceptorMock = $this->createMock(Interceptor::class);
+        $paymentInterceptorMock = $this->createMock(InfoInterface::class);
         $paymentInterceptorMock->expects($this->any())
             ->method('getAdditionalInformation')
             ->with(MpDeviceSessionId::MP_DEVICE_SESSION_ID)
             ->willReturn('arm:1234');
-        
+
         $paymentDataObjectMock = $this->createMock(PaymentDataObject::class);
         $paymentDataObjectMock->expects($this->once())
             ->method('getPayment')
             ->willReturn($paymentInterceptorMock);
-        
+
         $class = new MpDeviceSessionId();
 
         $result = $class->build(['payment' => $paymentDataObjectMock]);
@@ -28,19 +32,22 @@ class MpDeviceSessionIdTest extends TestCase {
         $this->assertEquals([MpDeviceSessionId::MP_DEVICE_SESSION_ID => 'arm:1234'], $result);
     }
 
+    /**
+     * @return void
+     */
     public function testBuildWithoutSession()
     {
-        $paymentInterceptorMock = $this->createMock(Interceptor::class);
+        $paymentInterceptorMock = $this->createMock(InfoInterface::class);
         $paymentInterceptorMock->expects($this->any())
             ->method('getAdditionalInformation')
             ->with(MpDeviceSessionId::MP_DEVICE_SESSION_ID)
             ->willReturn(null);
-        
+
         $paymentDataObjectMock = $this->createMock(PaymentDataObject::class);
         $paymentDataObjectMock->expects($this->once())
             ->method('getPayment')
             ->willReturn($paymentInterceptorMock);
-        
+
         $class = new MpDeviceSessionId();
 
         $result = $class->build(['payment' => $paymentDataObjectMock]);
@@ -48,10 +55,13 @@ class MpDeviceSessionIdTest extends TestCase {
         $this->assertEquals([], $result);
     }
 
+    /**
+     * @return void
+     */
     public function testBuildWithoutPayment()
     {
         $this->expectException(\InvalidArgumentException::class);
-        
+
         $class = new MpDeviceSessionId();
         $class->build([]);
     }